+++ lms/lms/doctype/loan/loan.py
@@ -205,6 +205,8 @@ class Loan(Document):
         transaction_id=None,
         loan_margin_shortfall_name=None,
         is_for_interest=None,
+        razorpay_event=None,
+        order_id=None,
     ):
         loan_transaction = frappe.get_doc(
             {
@@ -227,6 +229,10 @@ class Loan(Document):
             loan_transaction.loan_margin_shortfall = loan_margin_shortfall_name
         if is_for_interest:
             loan_transaction.is_for_interest = is_for_interest
+        if razorpay_event:
+            loan_transaction.razorpay_event = razorpay_event
+        if order_id:
+            loan_transaction.order_id = order_id
 
         loan_transaction.insert(ignore_permissions=True)